Some updates to the ref handling
Still getting a warning when running the App. It thinks that there is a mismatch between which hooks are being called during subsequent renderings. But I can't find the issue, and the page seems to be working, so I will leave it for now. Hoping to get rid of the warning in the future, since it's very annoying.